requirements.txt
fastapi>=0.110,<1.0
uvicorn[standard]>=0.27,<1.0
jinja2>=3.1,<4.0
httpx>=0.27,<1.0
feedparser>=6.0.11,<7.0
anthropic>=0.30,<1.0
python-dotenv>=1.0,<2.0
tenacity>=8.2,<9.0
flask>=3.0,<4.0
apscheduler>=3.10,<4.0
pytz>=2024.1
